Factor out `links` module in `h.views.api`
