+++ tests/h/views/api/annotations_test.py
@@ -5,40 +5,9 @@ import mock
 import pytest
 from webob.multidict import NestedMultiDict, MultiDict
 
-from pyramid import testing
-
 from h.schemas import ValidationError
 from h.search.core import SearchResult
 from h.views.api import annotations as views
-
-
-class TestLinks(object):
-
-    def test_it_returns_the_right_links(self, pyramid_config, pyramid_request):
-        pyramid_config.add_route('account', '/account/settings')
-        pyramid_config.add_route('forgot_password', '/forgot-password')
-        pyramid_config.add_route('group_create', '/groups/new')
-        pyramid_config.add_route('help', '/docs/help')
-        pyramid_config.add_route('oauth_authorize', '/oauth/authorize')
-        pyramid_config.add_route('oauth_revoke', '/oauth/revoke')
-        pyramid_config.add_route('activity.search', '/search')
-        pyramid_config.add_route('signup', '/signup')
-        pyramid_config.add_route('stream.user_query', '/u/{user}')
-
-        links = views.links(testing.DummyResource(), pyramid_request)
-
-        host = 'http://example.com'  # Pyramid's default host URL.
-        assert links == {
-            'account.settings': host + '/account/settings',
-            'forgot-password': host + '/forgot-password',
-            'groups.new': host + '/groups/new',
-            'help': host + '/docs/help',
-            'oauth.authorize': host + '/oauth/authorize',
-            'oauth.revoke': host + '/oauth/revoke',
-            'search.tag': host + '/search?q=tag:":tag"',
-            'signup': host + '/signup',
-            'user': host + '/u/:user',
-        }
 
 
 @pytest.mark.usefixtures('presentation_service', 'search_lib')
